backend/pyproject.toml
[tool.poetry]
name = "online-shop-backend"
version = "0.1.0"
description = ""
authors = ["Ahmad Nur <62164035+ahmaadn@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
aiofiles = "^23.2.1"
alembic = "^1.13.1"
bcrypt = "4.0.1"
email-validator = "^2.1.1"
fastapi = "^0.111.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
python = "^3.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.30"
uvicorn = "^0.29.0"
pymysql = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.2.0"
ruff = "^0.4.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
target_version = "py311"

[tool.ruff]
target-version = "py311"
# Same as Black.
line-length = 100

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "alembic",
  "frontend"
]

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
]

ignore = [
  "E501", # line too long, handled by black
  "B008", # do not perform function calls in argument defaults
  "W191", # indentation contains tabs
  "B904", # Allow raising exceptions without from e, for HTTPException
]